Implement HW9 1,2,3 Tasks

+++ Interactive_Menu/TelegramBot/UpdateHandler.cs
@@ -1,7 +1,8 @@
 ﻿using Interactive_Menu.Core.Entities;
 using Interactive_Menu.Core.Services;
-using Otus.ToDoList.ConsoleBot;
-using Otus.ToDoList.ConsoleBot.Types;
+using Telegram.Bot;
+using Telegram.Bot.Polling;
+using Telegram.Bot.Types;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -46,41 +47,51 @@ namespace Interactive_Menu.TelegramBot
         {
             try
             {
-                OnHandleEventStarted?.Invoke(update.Message.Text);
-                await botClient.SendMessage(update.Message.Chat, $"Получил '{update.Message.Text}'", ct);
+                if (update.Message is null) throw new ArgumentNullException();
+                if (update.Message.Text != null)
+                {
+                    OnHandleEventStarted?.Invoke(update.Message.Text);
+                    await botClient.SendMessage(update.Message.Chat, $"Получил '{update.Message.Text}'", cancellationToken: ct);
 
-                var command = update.Message.Text.Trim().ToLower(); // Получаем текст сообщения
-                var trimmedCommand = command.Split(' ', 2)[0];
+                    var command = update.Message.Text.Trim().ToLower(); // Получаем текст сообщения
+                    var trimmedCommand = command.Split(' ', 2)[0];
 
-                if (_commands.ContainsValue(trimmedCommand) && _isTaskCountLimitSet && _isTaskLengthLimitSet)
+                    if (_commands.ContainsValue(trimmedCommand) && _isTaskCountLimitSet && _isTaskLengthLimitSet)
+                    {
+                        await ExecuteCommand(botClient, update, trimmedCommand, ct); // Переходим к выполнению соответствующей команды
+                    }
+                    else if (!_isTaskCountLimitSet)
+                    {
+                        await SetTaskCountLimit(botClient, update, trimmedCommand, ct);
+                    }
+                    else if (!_isTaskLengthLimitSet)
+                    {
+                        await SetTaskLengthLimit(botClient, update, trimmedCommand, ct);
+                    }
+                    else
+                    {
+                        await botClient.SendMessage(update.Message.Chat, "Неизвестная команда.", cancellationToken: ct);
+                    }
+                    OnHandleEventCompleted?.Invoke(update.Message.Text);
+                } else
                 {
-                    await ExecuteCommand(botClient, update, trimmedCommand, ct); // Переходим к выполнению соответствующей команды
+                    await botClient.SendMessage(update.Message.Chat, "Получил пустой текст.", cancellationToken: ct);
                 }
-                else if (!_isTaskCountLimitSet)
-                {
-                    await SetTaskCountLimit(botClient, update, trimmedCommand, ct);
-                } else if (!_isTaskLengthLimitSet)
-                {
-                    await SetTaskLengthLimit(botClient, update, trimmedCommand, ct);
-                }
-                else
-                {
-                    await botClient.SendMessage(update.Message.Chat, "Неизвестная команда.", ct);
-                }
-                OnHandleEventCompleted?.Invoke(update.Message.Text);
+
             }
             catch (Exception Ex)
             {
-                await botClient.SendMessage(update.Message.Chat, "Произошла непредвиденная ошибка:", ct);
-                await botClient.SendMessage(update.Message.Chat, $"Тип исключения: {Ex.GetType()}", ct);
-                await botClient.SendMessage(update.Message.Chat, $"Исключение: {Ex.Message}", ct);
-                await botClient.SendMessage(update.Message.Chat, $"Трассировка стека: {Ex.StackTrace}", ct);
-                await botClient.SendMessage(update.Message.Chat, $"Внутреннее исключение: {Ex.InnerException}", ct);
+                await Console.Out.WriteLineAsync("Произошла непредвиденная ошибка:");
+                await Console.Out.WriteLineAsync($"Тип исключения: {Ex.GetType()}");
+                await Console.Out.WriteLineAsync($"Исключение: {Ex.Message}");
+                await Console.Out.WriteLineAsync($"Трассировка стека: {Ex.StackTrace}");
+                await Console.Out.WriteLineAsync($"Внутреннее исключение: {Ex.InnerException}");
             }
         }
 
         private async Task ExecuteCommand(ITelegramBotClient botClient, Update update, string command, CancellationToken ct)
         {
+            if (update.Message is null) throw new ArgumentNullException();
             if (_isAllCommandsAvailable)
             {
                 switch (command)
@@ -88,7 +99,7 @@ namespace Interactive_Menu.TelegramBot
                     case "/start": await OnStartCommand(botClient, update, ct); break;
                     case "/help": await OnHelpCommand(botClient, update, ct); break;
                     case "/info": await OnInfoCommand(botClient, update, ct); break;
-                    case "/exit": await OnExitCommand(botClient, update, ct); break;
+                    //case "/exit": await OnExitCommand(botClient, update, ct); break;
                     case "/addtask": await OnAddTaskCommand(botClient, update, ct); break;
                     case "/showtasks": await OnShowTasksCommand(botClient, update, ct); break;
                     case "/removetask": await OnRemoveTaskCommand(botClient, update, ct); break;
@@ -96,7 +107,7 @@ namespace Interactive_Menu.TelegramBot
                     case "/completetask": await OnCompleteTaskCommand(botClient, update, ct); break;
                     case "/report": await OnReportCommand(botClient, update, ct); break; 
                     case "/find": await OnFindCommand(botClient, update, ct); break;
-                    default: await botClient.SendMessage(update.Message.Chat, "Ошибка обработки команды.", ct); break;
+                    default: await botClient.SendMessage(update.Message.Chat, "Ошибка обработки команды.", cancellationToken: ct); break;
                 }
             } else
             {
@@ -105,8 +116,8 @@ namespace Interactive_Menu.TelegramBot
                     case "/start": await OnStartCommand(botClient, update, ct); break;
                     case "/help": await OnHelpCommand(botClient, update, ct); break;
                     case "/info": await OnInfoCommand(botClient, update, ct); break;
-                    case "/exit": await OnExitCommand(botClient, update, ct); break;
-                    default: await botClient.SendMessage(update.Message.Chat, "Ошибка обработки команды.", ct); break;
+                    //case "/exit": await OnExitCommand(botClient, update, ct); break;
+                    default: await botClient.SendMessage(update.Message.Chat, "Ошибка обработки команды.", cancellationToken: ct); break;
                 }
             }
         }
@@ -119,6 +130,7 @@ namespace Interactive_Menu.TelegramBot
         /// <param name="update"></param>
         private async Task OnFindCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.Text is null || update.Message.From is null) throw new ArgumentNullException();
             var task = update.Message.Text.Trim();
             string namePrefix = task.Remove(0, "/find".Length).Trim();
             var user = await _userService.GetUser(update.Message.From.Id, ct);
@@ -134,7 +146,7 @@ namespace Interactive_Menu.TelegramBot
                     for (int i = 0; i < tasksList.Count; i++)
                         outputBuilder.AppendLine($"Имя задачи {tasksList[i].Name} - {tasksList[i].CreatedAt} - {tasksList[i].Id}");
                 }
-                await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), ct);
+                await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), cancellationToken: ct);
             }
         }
 
@@ -146,28 +158,31 @@ namespace Interactive_Menu.TelegramBot
         /// <param name="update"></param>
         private async Task OnReportCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.From is null) throw new ArgumentNullException();
             var user = await _userService.GetUser(update.Message.From.Id, ct);
             if (user != null)
             {
                 var result = await _toDoReportService.GetUserStats(user.UserId, ct);
                 await botClient.SendMessage(update.Message.Chat,
                     $"Статистика по задачам на {result.generatedAt}. " +
-                    $"Всего: {result.total}; Завершенных: {result.completed}; Активных: {result.active};", ct);
+                    $"Всего: {result.total}; Завершенных: {result.completed}; Активных: {result.active};", cancellationToken: ct);
             }
         }
 
         private async Task OnCompleteTaskCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.Text is null) throw new ArgumentNullException();
             var task = update.Message.Text.Trim();
             task = task.Remove(0, "/completetask".Length).Trim();
             Guid taskId = new Guid(task);
             await _toDoService.MarkAsCompleted(taskId, ct);
 
-            await botClient.SendMessage(update.Message.Chat, $"Статус задачи {taskId} изменен", ct);
+            await botClient.SendMessage(update.Message.Chat, $"Статус задачи {taskId} изменен", cancellationToken: ct);
         }
 
         private async Task OnShowAllTasksCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.From is null) throw new ArgumentNullException();
             Guid userId = new Guid();
             var user = await _userService.GetUser(update.Message.From.Id, ct);
             if (user != null)
@@ -183,24 +198,24 @@ namespace Interactive_Menu.TelegramBot
                     for (int i = 0; i < tasksList.Count; i++)
                         outputBuilder.AppendLine($"({tasksList[i].State}) Имя задачи {tasksList[i].Name} - {tasksList[i].CreatedAt} - {tasksList[i].Id}");
                 }
-                await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), ct);
+                await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), cancellationToken: ct);
             }
-
         }
 
         private async Task OnRemoveTaskCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.Text is null) throw new ArgumentNullException();
             var task = update.Message.Text.Trim();
             task = task.Remove(0, "/removetask".Length).Trim();
             Guid taskId = new Guid(task);
             await _toDoService.Delete(taskId, ct);
-
-            await botClient.SendMessage(update.Message.Chat, $"Задача {taskId} удалена", ct);
+            await botClient.SendMessage(update.Message.Chat, $"Задача {taskId} удалена", cancellationToken: ct);
 
         }
 
         private async Task OnShowTasksCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.From is null) throw new ArgumentNullException();
             var user = await _userService.GetUser(update.Message.From.Id, ct);
             if (user != null)
             {
@@ -218,19 +233,20 @@ namespace Interactive_Menu.TelegramBot
                         if (tasksList[i].State == ToDoItemState.Active)
                             outputBuilder.AppendLine($"Имя задачи {tasksList[i].Name} - {tasksList[i].CreatedAt} - {tasksList[i].Id}");
                 }
-                await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), ct);
+                await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), cancellationToken: ct);
             }
         }
 
         private async Task OnAddTaskCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.Text is null || update.Message.From is null) throw new ArgumentNullException();
             var user = await _userService.GetUser(update.Message.From.Id, ct);
             var task = update.Message.Text.Trim();
             task = task.Remove(0, "/addtask".Length).Trim();
             if (user != null)
             {
                 await _toDoService.Add(user, task, ct);
-                await botClient.SendMessage(update.Message.Chat, $"Добавлена задача {task}", ct);
+                await botClient.SendMessage(update.Message.Chat, $"Добавлена задача {task}", cancellationToken: ct);
             }
         }
 
@@ -242,6 +258,7 @@ namespace Interactive_Menu.TelegramBot
 
         private async Task OnInfoCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null) throw new ArgumentNullException();
             StringBuilder outputBuilder = new StringBuilder();
             outputBuilder.AppendLine("\r\n" +
                                 "*  Текущая версия программы 7.0.  Дата создания 16-09-2025\r\n" +
@@ -259,11 +276,12 @@ namespace Interactive_Menu.TelegramBot
                                 "*  Версия программы 1.0.  Дата создания 25-05-2025\r\n" +
                                 "   Реализованы команды /start, /help, /info, /echo, /exit");
 
-            await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), ct);
+            await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), cancellationToken: ct);
         }
 
         private async Task OnHelpCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null) throw new ArgumentNullException();
             StringBuilder outputBuilder = new StringBuilder();
             outputBuilder.Append(
                 "Cправка по программе:" +
@@ -287,22 +305,23 @@ namespace Interactive_Menu.TelegramBot
                     "\r\n(например, /completetask 0167b785-b830-4d02-b82a-881b0b678034), программа завершает задачу, её статус становится Completed."
             );
 
-            await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), ct);
+            await botClient.SendMessage(update.Message.Chat, outputBuilder.ToString(), cancellationToken: ct);
         }
 
         private async Task OnStartCommand(ITelegramBotClient botClient, Update update, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.From is null) throw new ArgumentNullException();
             if (await _userService.GetUser(update.Message.From.Id, ct) != null)
             {
                 _isAllCommandsAvailable = true;
-                await botClient.SendMessage(update.Message.Chat, $"Привет, {update.Message.From.Username}", ct);
+                await botClient.SendMessage(update.Message.Chat, $"Привет, {update.Message.From.Username}", cancellationToken: ct);
             }
             else
             {
                 if (update.Message.From.Username != null)
                 {
                     await _userService.RegisterUser(update.Message.From.Id, update.Message.From.Username, ct);
-                    await botClient.SendMessage(update.Message.Chat, $"Привет, {update.Message.From.Username}", ct);
+                    await botClient.SendMessage(update.Message.Chat, $"Привет, {update.Message.From.Username}", cancellationToken: ct);
                     _isAllCommandsAvailable = true;
                 }
             }
@@ -313,7 +332,7 @@ namespace Interactive_Menu.TelegramBot
             {
                 _isTaskCountLimitSet = false;
                 await botClient.SendMessage(update.Message.Chat, $"{update.Message.From.Username}, " +
-                    $"введи максимальное количество задач от {_toDoService.MinTaskCountLimit} до {_toDoService.MaxTaskCountLimit}", ct);
+                    $"введи максимальное количество задач от {_toDoService.MinTaskCountLimit} до {_toDoService.MaxTaskCountLimit}", cancellationToken: ct);
             }    
         }
 
@@ -322,11 +341,12 @@ namespace Interactive_Menu.TelegramBot
         /// </summary>
         private async Task SetTaskCountLimit(ITelegramBotClient botClient, Update update, string command, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.From is null) throw new ArgumentNullException();
             _toDoService.TaskCountLimit = ParseAndValidateInt(command, _toDoService.MinTaskCountLimit, _toDoService.MaxTaskCountLimit);
             _isTaskCountLimitSet = true;
-            await botClient.SendMessage(update.Message.Chat, $"{update.Message.From.Username}, установлено максимальное количество задач: {_toDoService.TaskCountLimit}", ct);
+            await botClient.SendMessage(update.Message.Chat, $"{update.Message.From.Username}, установлено максимальное количество задач: {_toDoService.TaskCountLimit}", cancellationToken: ct);
 
-            await botClient.SendMessage(update.Message.Chat, $"{update.Message.From.Username}, введи максимальную длину задачи от {_toDoService.MinTaskLengthLimit} до {_toDoService.MaxTaskLengthLimit}", ct);
+            await botClient.SendMessage(update.Message.Chat, $"{update.Message.From.Username}, введи максимальную длину задачи от {_toDoService.MinTaskLengthLimit} до {_toDoService.MaxTaskLengthLimit}", cancellationToken: ct);
         }
 
         /// <summary>
@@ -334,9 +354,10 @@ namespace Interactive_Menu.TelegramBot
         /// </summary>
         private async Task SetTaskLengthLimit(ITelegramBotClient botClient, Update update, string command, CancellationToken ct)
         {
+            if (update.Message is null || update.Message.From is null) throw new ArgumentNullException();
             _toDoService.TaskLengthLimit = ParseAndValidateInt(command, _toDoService.MinTaskLengthLimit, _toDoService.MaxTaskLengthLimit);
             _isTaskLengthLimitSet = true;
-            await botClient.SendMessage(update.Message.Chat, $"{update.Message.From.Username}, установлена максимальная длина задачи: {_toDoService.TaskLengthLimit}", ct);
+            await botClient.SendMessage(update.Message.Chat, $"{update.Message.From.Username}, установлена максимальная длина задачи: {_toDoService.TaskLengthLimit}", cancellationToken: ct);
         }
 
         /// <summary>
@@ -358,9 +379,9 @@ namespace Interactive_Menu.TelegramBot
                 throw new ArgumentException($"Значение должно быть в диапазоне от {min} до {max}");
         }
 
-        public async Task HandleErrorAsync(ITelegramBotClient botClient, Exception exception, CancellationToken ct)
+        public async Task HandleErrorAsync(ITelegramBotClient botClient, Exception exception, HandleErrorSource source, CancellationToken cancellationToken)
         {
-            await Console.Out.WriteLineAsync($"Ошибка произошла: {exception.Message}");
+            await Console.Out.WriteLineAsync($"Ошибка произошла: {exception.Message}, источник: {source}");
         }
     }
 }
